requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.6
pyrebase4==4.7.1
requests==2.31.0
cryptography==42.0.5
